requirements.txt
flask
flask-cors
sqlalchemy
python-dotenv
PyJWT